requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-telegram-bot==20.6
requests==2.31.0